Add detected_at to vulnerabilities table
When working on
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/61385 we realized
there's no `detected_at` at column and `created_at` was used in it's
place (which made sense since up until now vulnerabilities were created
by scanners so `created_at == detected_at`).

Now that we want to create vulnerabilities via an API there can be cases
where `detected_at != created_at` and we want to be able to reflect
this.

In case `detected_at` is not provided it's assumed to be the same as
`created_at`.

Changelog: added