Remove securityScansSucceeded from DevOps Adoption
The securityScansSucceeded query was causing a timeout error. Also,
it is effectively being replaced with queries for individual scan
types. This removes the column from DevOps adoption, and removes the
data collection from the snapshot. This does not remove the column,
because this is an urgent change, so that will be removed in a future
chage.