Don't compress job arguments twice
When a job is scheduled in the future, it runs through the client
middleware twice. Once on scheduling, once when queueing.

Since the `args` key of the job would already be compressed at this
time, we don't need to compress it some more. Doing so would cause the
decompression to fail on the serverside.