Revise helptext on Datadog page
Bring this page a bit closer to GitLab tone and style. I've added a
commented-out section to provide an introductory paragraph. I copied
the formatting from another page, but I don't know enough about how
the code works, to make it work properly. I can ask an engineer to
review and polish it in a subsequent commit.