Migrate Edit Environments Form to Vue
Re-uses the new form component made as part of the migration of new
environments.

There was some question about whether or not we need Vuex or something,
but honestly because the form manages such a low number of properties
and the validation requirements are so small, that I opted to not use it
here instead.

Also the environment model is small, and so can be inlined in a data
attribute. As the validation is stricter on the backend, this should be
clear of HTML injection.