Use http read total timeout by default
Removes the use_read_total_timeout option on Gitlab::HTTP
and use read_total_timeout by default. The option was
introduced for HTTP calls to URLs that could be defined
by users in order to fix a security issue but have the
smallest possible impact. It looks like the timeout
implementation is stable so it can become the default
option. The only exception where We don't want to use
a read total timeout is when streaming potentially
large responses.